add MaybeIterableTypeTrait
