+++ src/Type/CallableType.php
@@ -3,11 +3,13 @@
 namespace PHPStan\Type;
 
 use PHPStan\TrinaryLogic;
+use PHPStan\Type\Traits\MaybeIterableTypeTrait;
 use PHPStan\Type\Traits\MaybeObjectTypeTrait;
 
 class CallableType implements CompoundType
 {
 
+	use MaybeIterableTypeTrait;
 	use MaybeObjectTypeTrait;
 
 	/**
@@ -45,21 +47,6 @@ class CallableType implements CompoundType
 	public function describe(): string
 	{
 		return 'callable';
-	}
-
-	public function isIterable(): TrinaryLogic
-	{
-		return TrinaryLogic::createMaybe();
-	}
-
-	public function getIterableKeyType(): Type
-	{
-		return new MixedType();
-	}
-
-	public function getIterableValueType(): Type
-	{
-		return new MixedType();
 	}
 
 	public function isCallable(): TrinaryLogic
